shared/rap: Add client ranging registration and notification parsing
Read the RAS Features characteristic to determine whether the remote
device supports real-time ranging. If supported, register for real-time
characteristic notifications using the reqtracker for the CS initiator
role.

Parse incoming segmented RAS ranging data notifications by accumulating
segments via iovec and parsing complete subevent headers and CS mode 0-3
step data, including IQ/tone PCT samples, once the last segment arrives.

+++ src/shared/rap.h
@@ -92,7 +92,7 @@ struct cs_mode_zero_data {
 	uint8_t packet_quality;
 	uint8_t packet_rssi_dbm;
 	uint8_t packet_ant;
-	uint32_t init_measured_freq_offset;
+	uint16_t init_measured_freq_offset;
 };
 
 struct cs_mode_one_data {
